async map in browser
