+++ lib/isThenable.js
@@ -1,0 +1,7 @@
+'use strict';
+
+function isThenable(obj) {
+	return (obj instanceof Object) && ('then' in obj) && typeof obj.then==="function";
+}
+
+module.exports = isThenable;
