+++ map.js
@@ -1,0 +1,6 @@
+"use strict";
+
+var Thenable = require('./lib/thenable');
+var isThenable = require('./lib/isThenable');
+
+module.exports = require('./covers/map')({Thenable: Thenable, isThenable: isThenable});
